requirements.txt
streamlit
crewai
litellm
groq
python-dotenv